requirements.txt
asgiref==3.7.2
astroid==2.15.6
dill==0.3.6
Django==4.2.3
djangorestframework==3.14.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
pep8==1.7.1
platformdirs==3.8.1
pylint==2.17.4
pytz==2023.3
sqlparse==0.4.4
tomlkit==0.11.8
typing_extensions==4.7.1
wrapt==1.15.0
requests==2.31.0
drf-yasg==1.21.6